backend/code/requirements.txt
async-timeout==4.0.2
bcrypt==3.2.0
beautifulsoup4==4.11.1
blinker==1.4
cachetools==5.2.0
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
Cmake
colorama==0.4.4
cryptography==3.4.8
dbus-python==1.2.18
defer==1.0.4
Deprecated==1.2.13
distro==1.7.0
distro-info===1.0
fasteners==0.14.1
Flask==2.2.2
Flask-Cors==3.0.10
future==0.18.2
google==3.0.0
google-api-core
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos
gunicorn==20.1.0
httplib2==0.20.2
idna
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==1.3
macaroonbakery
MarkupSafe==2.1.1
monotonic==1.6
more-itertools==8.10.0
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dateutil==2.8.1
python-dotenv==0.21.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==4.3.4
reportlab==3.6.8
requests
requests-oauthlib==1.3.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.2.post1
spotipy==2.20.0
uritemplate==4.1.1
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==2.2.2
wrapt==1.14.1
xdg
zipp==1.0.0
